# docs/10-client-api/Client-API.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EDUagent Client-API",
    "version": "5.0.0",
    "description": "前端 (Vue3) ↔ Backend (FastAPI, Port 8001) 接口规范。\n统一返回格式: {code, message, data}。\n流式接口使用 SSE (text/event-stream)。\n前端发起刷新/生成接口后收到 task_id，并通过 /tasks/{task_id} 轮询；Backend 负责创建任务、记录归属用户和任务类型；Agent Service 长任务完成后由 Backend 接收 webhook 回调，前端不直接等待 webhook。"
  },
  "servers": [
    {
      "url": "http://localhost:8001/api/v1",
      "description": "本地开发环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/admin/course-catalogs": {
      "get": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "管理员课程资源库列表",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCatalogListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "管理员创建课程资源库",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCatalogCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCatalogResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/course-catalogs/{catalog_id}": {
      "get": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "管理员课程资源库详情",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCatalogResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/course-catalogs/{catalog_id}/materials": {
      "get": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "课程资源库资料列表",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCatalogMaterialsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "登记课程资源库资料",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCatalogMaterialCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCatalogMaterialResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/course-catalogs/{catalog_id}/materials/{material_id}": {
      "delete": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "管理员软删除课程资源库资料",
        "description": "仅软删除资料记录，不删除上传文件，不删除 Qdrant chunks，不回滚 chunk_count。ready/partial 资源库删除资料后 knowledge_status 变为 dirty。",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "material_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "已删除",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCatalogMaterialDeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "资料不存在，业务码 40411"
          },
          "409": {
            "description": "课程资源库正在入库中"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/course-catalogs/{catalog_id}/materials/upload": {
      "post": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "上传课程资源库资料",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCatalogMaterialResponse"
                }
              }
            }
          },
          "400": {
            "description": "文件名、类型或内容非法"
          },
          "409": {
            "description": "课程资源库正在入库中"
          },
          "413": {
            "description": "资料文件过大"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/course-catalogs/{catalog_id}/ingestions": {
      "post": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "触发课程资源库知识入库",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "已接受",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCatalogIngestionResponse"
                }
              }
            }
          },
          "409": {
            "description": "正在入库或没有待入库资料"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/course-catalogs/{catalog_id}/knowledge-status": {
      "get": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "课程资源库知识库状态",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCatalogKnowledgeStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/course-catalogs/{catalog_id}/knowledge-graphs": {
      "get": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "管理员查看课程资源库知识图谱状态",
        "description": "Admin-only。返回 CourseCatalog 当前绑定 CourseOffering 的 course_id、active KG 摘要和最近一次 kg_generation 任务状态。没有 active KG 时 active_graph 为 null。",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCatalogKnowledgeGraphStatusResponse"
                }
              }
            }
          },
          "403": {
            "description": "非管理员访问"
          },
          "404": {
            "description": "课程资源库不存在"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/course-catalogs/{catalog_id}/knowledge-graphs/generations": {
      "post": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "管理员触发课程资源库知识图谱生成",
        "description": "Admin-only。默认基于 CourseCatalog 已入库知识切片自动创建新的 course_knowledge_graphs active 版本；请求体可为空，等价于 source_type=catalog_chunks, activate=true。kg_json 保留为后端调试兼容能力，Admin UI 不暴露。该接口不触发学生个性化 LearningPath 刷新。",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCatalogKnowledgeGraphGenerationRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "已接受",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCatalogKnowledgeGraphGenerationResponse"
                }
              }
            }
          },
          "403": {
            "description": "非管理员访问"
          },
          "404": {
            "description": "课程资源库不存在"
          },
          "409": {
            "description": "尚未绑定教学班或已有 KG 生成任务进行中；业务码 offering_missing / kg_task_running"
          },
          "422": {
            "description": "source_type 或输入内容不合法"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/course-catalogs/{catalog_id}/resources": {
      "get": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "管理员课程资源库生成资源列表",
        "description": "聚合该 CourseCatalog 下未软删除的共享资源，并兼容当前绑定教学班下尚未迁移的历史资源，供 Admin 在资源库抽屉中管理。",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCatalogResourcesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/course-catalogs/{catalog_id}/resources/generations": {
      "post": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "管理员触发课程资源库学习资源生成",
        "description": "Admin-only。基于 CourseCatalog 已入库知识生成标准学习资源。新生成资源归属当前 CourseCatalog，对所有绑定该资源库的教学班共享可读；当前仍兼容历史按教学班存储的旧资源，但新链路不再 fan-out 写多份资源。无绑定教学班时返回 409。",
        "parameters": [
          {
            "name": "catalog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogResourceGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "已接受",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogResourceGenerationResponse"
                }
              }
            }
          },
          "409": {
            "description": "资源库未就绪、知识库为空或尚未绑定教学班；业务码 40913/40914/40915"
          },
          "422": {
            "description": "资源类型为空或不合法，业务码 42210"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/resources/{resource_id}": {
      "delete": {
        "tags": [
          "Resources"
        ],
        "summary": "管理员软删除学习资源",
        "description": "仅设置 Resource.is_deleted=true，不删除数据库行、文件、Qdrant chunks 或 Agent 产物。",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "已删除",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResourceDeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在，业务码 40412"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/course-catalogs": {
      "get": {
        "tags": [
          "CourseCatalogs"
        ],
        "summary": "可绑定课程资源库列表",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "ready"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseCatalogReadyListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AdminCatalogKnowledgeGraphGenerationRequest": {
        "type": "object",
        "properties": {
          "source_type": {
            "type": "string",
            "enum": [
              "catalog_chunks",
              "outline_text",
              "kg_json"
            ],
            "default": "catalog_chunks",
            "description": "生成输入类型。默认 catalog_chunks，从资源库已入库知识切片自动生成 KG；kg_json 为后端调试兼容；outline_text 为历史兼容。"
          },
          "outline_text": {
            "type": "string",
            "nullable": true,
            "description": "source_type=outline_text 时必填，课程大纲文本；Admin UI 不暴露"
          },
          "kg_json": {
            "type": "object",
            "nullable": true,
            "description": "source_type=kg_json 时必填，包含 nodes 和 edges 的知识图谱 JSON；后端调试兼容能力，Admin UI 不暴露"
          },
          "activate": {
            "type": "boolean",
            "default": true,
            "description": "生成后是否设为 active 版本"
          }
        }
      },
      "AdminCatalogKnowledgeGraphGenerationResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 202
          },
          "message": {
            "type": "string",
            "example": "accepted"
          },
          "data": {
            "type": "object",
            "properties": {
              "task_id": {
                "type": "string"
              },
              "catalog_id": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "example": "processing"
              }
            }
          }
        }
      },
      "AdminCatalogKnowledgeGraphStatusResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/AdminCatalogKnowledgeGraphStatusData"
          }
        }
      },
...
      "CourseCatalogCreateRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "CourseCatalogResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CourseCatalogItem"
          }
        }
      },
      "CourseCatalogListResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "catalogs": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CourseCatalogItem"
                }
              },
              "total": {
                "type": "integer"
              },
              "page": {
                "type": "integer"
              },
              "page_size": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "CourseCatalogMaterialCreateRequest": {
        "type": "object",
        "required": [
          "filename",
          "source_type"
        ],
        "properties": {
          "filename": {
            "type": "string"
          },
          "source_type": {
            "type": "string"
          },
          "storage_uri": {
            "type": "string"
          }
        }
      },
...
      "CourseCatalogIngestionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CourseCatalogIngestionAccepted"
          }
        }
      },
      "CourseCatalogMaterialResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CourseCatalogMaterialItem"
          }
        }
      },
      "CourseCatalogMaterialsResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "materials": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CourseCatalogMaterialItem"
                }
              }
            }
          }
        }
      },
...
      "CourseCatalogMaterialDeleteResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "catalog_id": {
                "type": "string"
              },
              "deleted": {
                "type": "boolean"
              },
              "knowledge_status": {
                "type": "string",
                "enum": [
                  "draft",
                  "ingesting",
                  "ready",
                  "dirty",
                  "partial",
                  "failed"
                ]
              }
            }
          }
        }
      },
      "CourseCatalogKnowledgeStatusResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "catalog_id": {
                "type": "string"
              },
              "status": {
                "type": "string"
              },
              "knowledge_status": {
                "type": "string",
                "enum": [
                  "draft",
                  "ingesting",
                  "ready",
                  "dirty",
                  "partial",
                  "failed"
                ]
              },
              "material_count": {
                "type": "integer"
              },
              "chunk_count": {
                "type": "integer"
              },
              "pending_material_count": {
                "type": "integer"
              },
              "failed_material_count": {
                "type": "integer"
              },
              "last_ingestion_task_id": {
                "type": "string",
                "nullable": true
              },
              "last_ingestion_status": {
                "type": "string",
                "nullable": true
              },
              "last_error": {
                "type": "string",
                "nullable": true
              }
            }
          }
        }
      },
...
    }
  }
}